requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
python-dotenv>=1.0.1
groq>=0.4.2
nest-asyncio>=1.6.0
langchain>=0.1.10
langchain-community>=0.0.25
langchain-groq>=0.0.1
langchain-ollama>=0.0.1
langchain-huggingface>=0.0.1
transformers>=4.38.0
torch>=2.2.0
requests>=2.31.0
yfinance>=0.2.36
faiss-cpu>=1.8.0
sentence-transformers>=2.5.1
google-search-results>=2.4.2
